Slim down the update action.
Delete all the fields--because order shouldn't be preserved and everything about the field could have changed.
I also see no need to allow two different save actions on this page; it is possible to allow for the same
workflow while simplifying things a tiny bit. Put errors in the flash (for now).